requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
python-dotenv==0.13.0
requests==2.23.0
soupsieve==2.0
urllib3==1.25.9
